egl2 branch: Always use float vertex positions because the resolution isn't high enough otherwise.
svn path=/branches/libavg_egl2/; revision=8411

+++ src/graphics/VertexData.h
@@ -41,7 +41,7 @@ struct FloatVertex {
 
 struct FixedVertex {
     GLshort m_Tex[2];    // 4.12 fixed point.
-    GLshort m_Pos[2];    // 12.4 fixed point.
+    GLfloat m_Pos[2];
     Pixel32 m_Color;
 };
